+++ src/main/java/com/maiya/bean/WebUser.java
@@ -19,6 +19,8 @@ public class WebUser implements Serializable {
 
     private Byte assess;
 
+    private User user;
+
     private static final long serialVersionUID = 1L;
 
     public Long getId() {
@@ -139,5 +141,13 @@ public class WebUser implements Serializable {
         sb.append(", serialVersionUID=").append(serialVersionUID);
         sb.append("]");
         return sb.toString();
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
     }
 }
