--- src/main/java/com/maiya/impl/UserServiceImpl.java
@@ -1,109 +1,0 @@
-package com.maiya.impl;
-
-import com.maiya.bean.User;
-import com.maiya.dao.UserMapper;
-import com.maiya.service.UserService;
-import org.springframework.stereotype.Service;
-
-import javax.annotation.Resource;
-import java.util.HashMap;
-import java.util.List;
-
-@Service("userService")
-public class UserServiceImpl implements UserService {
-	@Resource  
-    private UserMapper userDao;
-
-	public User selectWebUserByPassword(String name, String password) {
-		return userDao.selectWebUserByPassword(name, password);
-	}
-
-    public User selectAppUserByPassword(String name, String password) {
-        return userDao.selectAppUserByPassword(name, password);
-    }
-	
-	/**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table user_tbl
-     *
-     * @mbggenerated
-     */
-    public int deleteByPrimaryKey(Long id) {
-    	return userDao.deleteByPrimaryKey(id);
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table user_tbl
-     *
-     * @mbggenerated
-     */
-    public int insertSelective(User record) {
-    	return userDao.insertSelective(record);
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table user_tbl
-     *
-     * @mbggenerated
-     */
-    public User selectByPrimaryKey(Long id) {
-    	return userDao.selectByPrimaryKey(id);
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table user_tbl
-     *
-     * @mbggenerated
-     */
-    public int updateByPrimaryKeySelective(User record) {
-    	return userDao.updateByPrimaryKeySelective(record);
-    }
-
-    public List<Object> selectAppAllUser(String name) {
-        return userDao.selectAppAllUser(name);
-    }
-
-    public List<Object> selectAllAppByUserId(HashMap<String, Object> parameters) {
-        return userDao.selectAllAppByUserId(parameters);
-    }
-
-    public List<Object> selectWebAllUser(String name) {
-        return userDao.selectWebAllUser(name);
-    }
-
-    public List<Object> selectAllWebByUserId(HashMap<String, Object> parameters) {
-        return userDao.selectAllWebByUserId(parameters);
-    }
-
-    public User selectVerifyCodeByPhone(String phone_number) {
-        return userDao.selectVerifyCodeByPhone(phone_number);
-    }
-
-    public int updateByPhone(User record) {
-        return userDao.updateByPhone(record);
-    }
-
-    public Long selectLastUserId() {
-        return userDao.selectLastUserId();
-    }
-
-    public User selectWebUserByUserId(Long id) {
-        return userDao.selectWebUserByUserId(id);
-    }
-
-    public User selectAppUserByUserId(Long id) {
-        return userDao.selectAppUserByUserId(id);
-    }
-
-    public int deleteWebByPrimaryKey(Long id) {
-        return userDao.deleteWebByPrimaryKey(id);
-    }
-
-    public int deleteAppByPrimaryKey(Long id) {
-        return userDao.deleteAppByPrimaryKey(id);
-    }
-}
-
