+++ src/main/java/com/maiya/mapper/UserMapper.xml
@@ -40,132 +40,6 @@
     id_card
   </sql>
 
-  <select id="selectLastUserId" resultMap="BaseResultMap">
-    select max(id) from user_tbl
-  </select>
-
-  <select id="selectWebUserByUserId" resultMap="BaseResultMap" parameterType="java.lang.Long">
-    select
-      user_tbl.*,
-      web_user_tbl.*
-    from user_tbl
-    <!--LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id-->
-      LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
-    where 1=1
-      and user_tbl.id = #{id,jdbcType=BIGINT}
-  </select>
-
-  <select id="selectAppUserByUserId" resultMap="BaseResultMap" parameterType="java.lang.Long">
-    select
-    user_tbl.*,
-    app_user_tbl.*
-    from user_tbl
-    <!--LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id-->
-    LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id
-    where 1=1
-    and user_tbl.id = #{id,jdbcType=BIGINT}
-  </select>
-
-  <select id="selectWebUserByPassword" resultMap="BaseResultMap" >
-    <!--
-      WARNING - @mbggenerated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    select
-      user_tbl.*,
-      web_user_tbl.*
-    from user_tbl
-      <!--LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id-->
-      LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
-    where 1=1
-      and user_tbl.name = #{0}
-      and user_tbl.password = #{1}
-  </select>
-
-  <select id="selectAppUserByPassword" resultMap="BaseResultMap" >
-    <!--
-      WARNING - @mbggenerated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    select
-    user_tbl.*,
-    app_user_tbl.*
-    from user_tbl
-    LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id
-    where 1=1
-    and user_tbl.name = #{0}
-    and user_tbl.password = #{1}
-  </select>
-
-  <select id="selectAppAllUser" resultMap="BaseResultMap">
-    select
-      user_tbl.*,
-      app_user_tbl.*
-    from user_tbl
-    LEFT JOIN app_user_tbl ON  app_user_tbl.user_id = user_tbl.id
-    where 1 = 1
-    <if test="_parameter != null and _parameter != ''">
-      AND user_tbl.name like CONCAT('%',#{_parameter,jdbcType=VARCHAR},'%')
-    </if>
-    ORDER BY app_user_tbl.id
-  </select>
-
-  <select id="selectAllAppByUserId" resultMap="BaseResultMap" parameterType="hashmap">
-    <!--
-      WARNING - @mbggenerated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    select
-      user_tbl.*,
-      app_user_tbl.*
-    from  user_tbl
-    LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id
-    where 1 = 1
-    <if test="name != null and name != ''">
-      AND user_tbl.name like CONCAT('%',#{name,jdbcType=VARCHAR},'%')
-    </if>
-    <if test="userId != null">
-      AND user_tbl.id = #{id,jdbcType=BIGINT}
-    </if>
-    ORDER BY app_user_tbl.id
-  </select>
-
-  <select id="selectWebAllUser" resultMap="BaseResultMap" parameterType="java.lang.String">
-    <!--
-      WARNING - @mbggenerated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    select
-    user_tbl.*,
-    web_user_tbl.*
-    from user_tbl
-    LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
-    where 1 = 1
-    <if test="_parameter != null and _parameter != ''">
-      AND user_tbl.name like CONCAT('%',#{_parameter,jdbcType=VARCHAR},'%')
-    </if>
-    ORDER BY web_user_tbl.id
-  </select>
-
-  <select id="selectAllWebByUserId" resultMap="BaseResultMap" parameterType="hashmap">
-    <!--
-      WARNING - @mbggenerated
-      This element is automatically generated by MyBatis Generator, do not modify.
-    -->
-    select
-    user_tbl.*,
-    web_user_tbl.*
-    from  user_tbl
-    LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
-    where 1 = 1
-    <if test="name != null and name != ''">
-      AND user_tbl.name like CONCAT('%',#{name,jdbcType=VARCHAR},'%')
-    </if>
-    <if test="userId != null">
-      AND user_tbl.id = #{id,jdbcType=BIGINT}
-    </if>
-    ORDER BY web_user_tbl.id
-  </select>
 
   <select id="selectVerifyCodeByPhone" resultMap="BaseResultMap" parameterType="string">
     <!--
@@ -249,12 +123,6 @@
     where id = #{id,jdbcType=BIGINT}
   </delete>
 
-  <delete id="deleteWebByPrimaryKey" parameterType="java.lang.Long">
-    delete user_tbl, web_user_tbl from user_tbl, web_user_tbl
-    where
-      web_user_tbl.user_id = user_tbl.id
-      AND user_tbl.id = #{id,jdbcType=BIGINT}
-  </delete>
 
   <delete id="deleteAppByPrimaryKey" parameterType="java.lang.Long">
     delete user_tbl, app_user_tbl from user_tbl, app_user_tbl
@@ -358,6 +226,7 @@
       </if>
     </trim>
   </insert>
+
   <update id="updateByPrimaryKeySelective" parameterType="com.maiya.bean.User">
     update user_tbl
     <set>
@@ -416,43 +285,4 @@
     where id = #{id,jdbcType=BIGINT}
   </update>
 
-  <update id="updateWebByPrimaryKey" parameterType="com.maiya.bean.User">
-    update user_tbl as u, web_user_tbl as w
-    set u.name = #{name,jdbcType=VARCHAR},
-    u.sex = #{sex,jdbcType=TINYINT},
-    u.age = #{age,jdbcType=TINYINT},
-    u.avatar = #{avatar,jdbcType=VARCHAR},
-    u.email = #{email,jdbcType=VARCHAR},
-    u.phone_number = #{phoneNumber,jdbcType=VARCHAR},
-    u.verification = #{verification,jdbcType=VARCHAR},
-    u.password = #{password,jdbcType=VARCHAR},
-    u.background_icon = #{backgroundIcon,jdbcType=VARCHAR},
-    u.id_card = #{idCard,jdbcType=VARCHAR},
-    w.careers = #{careers,jdbcType=VARCHAR},
-    w.privilege = #{privilege,jdbcType=VARCHAR},
-    w.privilegetype = #{privilegetype,jdbcType=TINYINT},
-    w.certificate = #{certificate,jdbcType=VARCHAR},
-    w.identity = #{identity,jdbcType=BIT},
-    w.assess = #{assess,jdbcType=TINYINT}
-    where u.id = w.user_id
-    AND id = #{id,jdbcType=BIGINT}
-  </update>
-
-  <update id="updateAppByPrimaryKey" parameterType="com.maiya.bean.User">
-    update user_tbl as u, app_user_tbl as a
-    set u.name = #{name,jdbcType=VARCHAR},
-    u.sex = #{sex,jdbcType=TINYINT},
-    u.age = #{age,jdbcType=TINYINT},
-    u.avatar = #{avatar,jdbcType=VARCHAR},
-    u.email = #{email,jdbcType=VARCHAR},
-    u.phone_number = #{phoneNumber,jdbcType=VARCHAR},
-    u.verification = #{verification,jdbcType=VARCHAR},
-    u.password = #{password,jdbcType=VARCHAR},
-    u.background_icon = #{backgroundIcon,jdbcType=VARCHAR},
-    u.id_card = #{idCard,jdbcType=VARCHAR},
-    a.address = #{address,jdbcType=VARCHAR},
-    a.points = #{points,jdbcType=SMALLINT}
-    where u.id = a.user_id
-    AND u.id = #{id,jdbcType=BIGINT}
-  </update>
 </mapper>
