+++ src/main/java/com/maiya/mapper/WebUserMapper.xml
@@ -10,10 +10,126 @@
     <result column="certificate" jdbcType="VARCHAR" property="certificate" />
     <result column="identity" jdbcType="BIT" property="identity" />
     <result column="assess" jdbcType="TINYINT" property="assess" />
+    <collection property="user" resultMap="userResult" />
   </resultMap>
+
+  <resultMap id="userResult" type="com.maiya.bean.User">
+    <id column="id" jdbcType="BIGINT" property="id" />
+    <result column="name" jdbcType="VARCHAR" property="name" />
+    <result column="sex" jdbcType="TINYINT" property="sex" />
+    <result column="age" jdbcType="TINYINT" property="age" />
+    <result column="avatar" jdbcType="VARCHAR" property="avatar" />
+    <result column="email" jdbcType="VARCHAR" property="email" />
+    <result column="phone_number" jdbcType="VARCHAR" property="phoneNumber" />
+    <result column="verification" jdbcType="VARCHAR" property="verification" />
+    <result column="password" jdbcType="VARCHAR" property="password" />
+    <result column="background_icon" jdbcType="VARCHAR" property="backgroundIcon" />
+    <result column="id_card" jdbcType="VARCHAR" property="idCard" />
+  </resultMap>
+
   <sql id="Base_Column_List">
     id, user_id, careers, privilege, privilegetype, certificate, identity, assess
   </sql>
+
+  <select id="selectLastUserId" resultMap="BaseResultMap">
+    select max(id) from user_tbl
+  </select>
+
+  <select id="selectWebUserByUserId" resultMap="BaseResultMap" parameterType="java.lang.Long">
+    select
+    user_tbl.*,
+    web_user_tbl.*
+    from web_user_tbl
+    LEFT JOIN user_tbl ON user_tbl.id = web_user_tbl.user_id
+    where 1=1
+    and user_tbl.id = #{id,jdbcType=BIGINT}
+  </select>
+
+  <select id="selectWebUserByPassword" resultMap="BaseResultMap" >
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    select
+    user_tbl.*,
+    web_user_tbl.*
+    from user_tbl
+    <!--LEFT JOIN app_user_tbl ON user_tbl.id = app_user_tbl.user_id-->
+    LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
+    where 1=1
+    and user_tbl.name = #{0}
+    and user_tbl.password = #{1}
+  </select>
+
+
+  <select id="selectWebAllUser" resultMap="BaseResultMap" parameterType="java.lang.String">
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    select
+    user_tbl.*,
+    web_user_tbl.*
+    from user_tbl
+    LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
+    where 1 = 1
+    <if test="_parameter != null and _parameter != ''">
+      AND user_tbl.name like CONCAT('%',#{_parameter,jdbcType=VARCHAR},'%')
+    </if>
+      AND web_user_tbl.user_id > 0
+    ORDER BY web_user_tbl.id
+  </select>
+
+  <select id="selectAllWebByUserId" resultMap="BaseResultMap" parameterType="hashmap">
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    select
+    user_tbl.*,
+    web_user_tbl.*
+    from  user_tbl
+    LEFT JOIN web_user_tbl ON user_tbl.id = web_user_tbl.user_id
+    where 1 = 1
+    <if test="name != null and name != ''">
+      AND user_tbl.name like CONCAT('%',#{name,jdbcType=VARCHAR},'%')
+    </if>
+    <if test="userId != null">
+      AND user_tbl.id = #{id,jdbcType=BIGINT}
+    </if>
+    ORDER BY web_user_tbl.id
+  </select>
+
+  <delete id="deleteWebByPrimaryKey" parameterType="java.lang.Long">
+    delete user_tbl, web_user_tbl from user_tbl, web_user_tbl
+    where
+    web_user_tbl.user_id = user_tbl.id
+    AND user_tbl.id = #{id,jdbcType=BIGINT}
+  </delete>
+
+
+  <update id="updateWebByPrimaryKey" parameterType="com.maiya.bean.User">
+    update user_tbl as u, web_user_tbl as w
+    set u.name = #{name,jdbcType=VARCHAR},
+    u.sex = #{sex,jdbcType=TINYINT},
+    u.age = #{age,jdbcType=TINYINT},
+    u.avatar = #{avatar,jdbcType=VARCHAR},
+    u.email = #{email,jdbcType=VARCHAR},
+    u.phone_number = #{phoneNumber,jdbcType=VARCHAR},
+    u.verification = #{verification,jdbcType=VARCHAR},
+    u.password = #{password,jdbcType=VARCHAR},
+    u.background_icon = #{backgroundIcon,jdbcType=VARCHAR},
+    u.id_card = #{idCard,jdbcType=VARCHAR},
+    w.careers = #{careers,jdbcType=VARCHAR},
+    w.privilege = #{privilege,jdbcType=VARCHAR},
+    w.privilegetype = #{privilegetype,jdbcType=TINYINT},
+    w.certificate = #{certificate,jdbcType=VARCHAR},
+    w.identity = #{identity,jdbcType=BIT},
+    w.assess = #{assess,jdbcType=TINYINT}
+    where u.id = w.user_id
+    AND id = #{id,jdbcType=BIGINT}
+  </update>
+
   <select id="selectByPrimaryKey" parameterType="java.lang.Long" resultMap="BaseResultMap">
     select 
     <include refid="Base_Column_List" />
