+++ src/main/java/com/maiya/service/UserService.java
@@ -1,6 +1,7 @@
 package com.maiya.service;
 
 import com.maiya.bean.User;
+import com.maiya.bean.WebUser;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.HashMap;
@@ -39,29 +40,13 @@ public interface UserService {
      */
     public int updateByPrimaryKeySelective(User record);
 
-    public User selectWebUserByPassword(@Param("name") String name, @Param("password") String password);
-
     public User selectAppUserByPassword(@Param("name") String name, @Param("password") String password);
 
 	public List<Object> selectAppAllUser(String name);
 
-    public List<Object> selectAllAppByUserId(HashMap<String, Object> parameters);
-
-    public List<Object> selectWebAllUser(String name);
-
-    public List<Object> selectAllWebByUserId(HashMap<String, Object> parameters);
-
-    public User selectVerifyCodeByPhone(String phone_number);
-
-    public int updateByPhone(User record);
-
     public Long selectLastUserId();
 
-    public User selectWebUserByUserId(Long id);
-
     public User selectAppUserByUserId(Long id);
-
-    public int deleteWebByPrimaryKey(Long id);
 
     public int deleteAppByPrimaryKey(Long id);
 }
